PROCESSING-3868: Moving logic to http module. It does not add anything keeping it as a separated class
